Matchstick_Game.py
The Matchstick Game is a game for two players. A player chooses one or two matchstick in a turn.
The player to choose the last matchstick is declared a winner.
This program gives an option to play either against a computer or against a second player.
The program uses basic concepts of Python like while, if-else, input and string matching.
The problem solution was written by me.
Error capturing can be added to the program to make it more professional.