The Hangman Word Game
Hangman is a popular word guessing game.
The player is can choose the category of the words to guess - to make the game simpler.
The player is given 15 chances (it can be reduced to 10 if required)
The problem solution uses the concepts of strings. If you are referring to the py4e website, this program covers all concepts till chapter 6.
However, the 'random' library is imported.
Importing libraries is an advanced concept (if you are referring to the py4e website).
In that case, a short cut is to allow the player to enter  a random number between 1 to 12. (Code for this is hidden. Keyword is r_number. Error Capturing for this section is not done.)