Use the middle letter to make words.
This program uses a tuple to display a shuffled word as the puzzle word.
The rest of is basically a copy of the "make_words" game. However, the player is not given clues on how many words are available as per the word length. No bonus points are given either.
--- GAME CONCEPT ----
#The game is based on the game that appears daily in Times of India
#The orginal game has a center letter surrounded by six other letters in a hexagon
#The player has to make words using more than one letter as well as the center letter.
#Words that don't include the center letter don't qualify as a valid answer.
#One seven letter word has to be formed.
#In this game, the computer generates the word for the puzzle
#If a player spells more than half of the correct answers, he is considered a winner.